Setup README.md and HTML file
This commit sets the first iteration of the html document, with the
header section, middle sections, sign up section, and footer section
divided.
Each section either has anchor tags as placeholders, with the pseudo
text provided to follow the design as well as for CSS styling.
Or a button, as well.