Create checklist markdown file and begin styling
In this commit a checklist.md file was created just for ones own sake
of keeping track of current progress. Styles have began being added to
the page once the first iteration of html was made. Changes to
the HTML (dividing areas into new containers) were made. This is just
as a result of the learning curve, removing things where not necessary,
or trying new things and making discoveries.
The styles applied to the header area, have reached a desirable state
at this point. Enough to carry onto the top middle middle, which I began
working on and applying styles to the text, and working with flex
to get it positioned correctly.