Revert project to an earlier commit state
This project was reverted to the state where the HTML was just
completed, but the CSS was not started.
The reason for this is because I felt that I overcomplicated things,
not having a fair enough understanding of CSS. So with some more
practice, and newfound knowledge, I will be attempting to tackle it
from scratch again, and see how the approach may vary this time around.