Finish header and edit HTML document
The header is for the moment in ones eyes complete, in positioning,
and in styling. The HTML document was edited to take away any unwanted
margin from the body, and it started to look more like how the finished
product would. A wrapper added around everything, which may be
necessary for everything to be properly spaced on the screen.
We will see. The HTML may be changed again but for the moment how it
is is okay. Next commit will be working with the Hero Main Section.