Add styling to top main section and edit HTML
More HTML was either removed or changed, for sake of having things
properly contained for flex, or because they simply were not in the
correct order, causes issues. Like the button being switched with the
elements that contained the placeholder text. Besides this,
styling has been adding for the buttons and the text, though this
section is still incomplete.