Continued styling the top header section
Color and changes to font size were done to the header logo, the header
link texts, color change to the text themselves, as well as color to the
background behind the text. There was a slight adjustment to the HTML,
where the <nav> tag was removed that was around the top section, and
instead left it as a div wrapping all the elements in the header logo
section.