Styled information section and edited HTML
I edited the HTML document to wrap a div container named information-
item to work on an issue I was having of not being able to put the text
under the placeholder image boxes. It ended up working. And then after
that I continued to style more of the information section. Besides
maybe working on the margin a little more, to put the boxes a little
closer together and to the center, I will be finished with that section.
And then will begin to work on the quote section.