Finished signup section styles and edited HTML
The HTML document was edited so that the button in the signup section
was instead outside of the div named "call-to-action-container", so
that we can apply flex and be able to place the button to the right of
the text inside the box. Maybe there is a better way to do this, but
this is the way that I came up with eventually.
And on top of that, the styling was completed for that section as well.
Now what is left before the landing page is complete is to work on the
footer.