Removed footer class and edited quote container
In the quote container I made a slight change to the top and bottom
margin so that the container is tighter together on the top and bottom.
Other than that, inside the HTML document, I edited the footer to
remove an unnecessary class name. For what we are going to do, the
footer tag will suffice without the class name.